src/requirements.txt
langchain>=0.1.1
langchain-community
langchainhub
pypdf
gpt4all
streamlit
pinecone-client
langchain-pinecone
google-generativeai

